requirements-dev.txt
audata
bcrypt
cython
email_validator
flask
flask_login
flask_mail
flask_socketio
flask_sqlalchemy
flask_wtf
htmlmin
numpy
pandas
passlib
psutil
pycrypto
simplejson
sqlalchemy